backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
drf-yasg==1.21.7
celery==5.3.6
redis==5.0.1
pillow==10.2.0
django-storages==1.14.2
pytest==7.4.4
pytest-django==4.7.0
coverage==7.4.1
black==23.12.1
flake8==7.0.0 